Fix/ code refacto into components + Fix bug wrap text into cards
