+++ app/javascript/app.vue
@@ -1,58 +1,25 @@
 <template>
   <draggable v-model="lists" v-bind="{group: 'lists'}" class="board dragArea" @end="listMoved">
-    <div v-for="(list, index) in original_lists" class="list">
-      <h6>{{ list.name }}</h6>
-
-      <draggable v-model="list.cards" v-bind="{group: 'cards'}" class="dragArea" @change="cardMoved">
-        <div v-for="(card, index) in list.cards" class="card card-body">
-          {{ card.name }}
-        </div>
-      </draggable>
-
-      <textarea v-model="messages[list.id]" class="form-control mb-1"></textarea>
-      <button v-on:click="submitMessages(list.id)" class="btn btn-primary"">Add</button>
-    </div>
+    <list v-for="(list, index) in lists" :list="list"></list>
   </draggable>
 </template>
 
 <script>
 import Rails from '@rails/ujs';
 import draggable from 'vuedraggable';
+import list from 'components/list';
 
   export default {
-    components: { draggable },
+    components: { draggable, list },
 
     props: ["original_lists"],
     data: function() {
       return {
-        messages: {},
         lists: this.original_lists,
       }
     },
 
     methods: {
-      cardMoved: function(event){
-        const evt = event.added || event.moved
-        if (evt == undefined) { return }
-
-        const element = evt.element
-        const list_index = this.lists.findIndex((list) => {
-          return list.cards.find((card) => {
-            return card.id === element.id
-          })
-        })
-
-        var data = new FormData
-        data.append("card[list_id]", this.lists[list_index].id)
-        data.append("card[position]", evt.newIndex + 1)
-
-        Rails.ajax({
-          url: `cards/${element.id}/move`,
-          type: "PATCH",
-          data: data,
-          dataType: "json",
-        })
-      },
 
       listMoved: function(event) {
         var data = new FormData
@@ -65,24 +32,6 @@ import draggable from 'vuedraggable';
           dataType: "json",
         })
       },
-
-      submitMessages: function(list_id) {
-        var data = new FormData
-        data.append("card[list_id]", list_id)
-        data.append("card[name]", this.messages[list_id])
-
-        Rails.ajax({
-          url: "/cards",
-          type: "POST",
-          data: data,
-          dataType: "json",
-          success: (data) => {
-            const index = this.lists.findIndex(item => item.id == list_id)
-            this.lists[index].cards.push(data)
-            this.messages[list_id] = undefined
-          }
-        })
-      }
     }
   }
 
@@ -100,17 +49,7 @@ import draggable from 'vuedraggable';
   }
 
   .board {
-    white-space: nowrap;
+    white-space: wrap;
     overflow-x: auto;
-  }
-
-  .list{
-    display: inline-block;
-    background: #F4FAFF;
-    border-radius: 8px;
-    margin-right: 20px;
-    padding: 20px;
-    vertical-align: top;
-    width: 270px;
   }
 </style>
